+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## [0.0.6-beta] - 2024-xx-xx
 ### Added
 - rule based gate reduction
+- merging rotation using phase polynomials and propagation of x gates
 
 ## [0.0.5] - 2023-12-23
 ### Changed